requirements.txt
google-generativeai>=0.8.3
chromadb>=0.5.0
pypdf2>=3.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
streamlit>=1.30.0
tabulate>=0.9.0